HasLocale util package was locale
- Necessary to avoid split packages with walkingkooka-locale
