--- src/main/java/walkingkooka/locale/HasLocaleTesting.java
@@ -1,34 +1,0 @@
-/*
- * Copyright 2019 Miroslav Pokorny (github.com/mP1)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-package walkingkooka.locale;
-
-import walkingkooka.test.Testing;
-
-import java.util.Locale;
-
-public interface HasLocaleTesting extends Testing {
-
-    default void localeAndCheck(final HasLocale has,
-                                final Locale expected) {
-        this.checkEquals(
-            expected,
-            has.locale(),
-            () -> has + " locale()"
-        );
-    }
-}
